Convert to Python 3

+++ centos/centos_cert.py
@@ -21,7 +21,7 @@ class CentOSUserCert(object):
             # are all pieces of data we want to reference in this class
             self.__dict__.update(dict(self._cert.get_subject().get_components()))
 
-            self.expired = self._cert.has_expired() != 0L
+            self.expired = self._cert.has_expired() != 0
             self.serial = self._cert.get_serial_number()
 
     @property
@@ -29,7 +29,7 @@ class CentOSUserCert(object):
         crl_response = requests.get(defaults.FAS_CRL)
         crl = crypto.load_crl(crypto.FILETYPE_PEM, crl_response.text)
 
-        if self.serial in (long(x.get_serial(), 16) for x in crl.get_revoked()):
+        if self.serial in (int(x.get_serial(), 16) for x in crl.get_revoked()):
             return False
 
         if self.expired:
